connection front and back for storage and product

+++ CommercialApplication/ApplicationLayer/Dtoes/Product/ProductStorageDto.cs
@@ -2,8 +2,8 @@
 {
     public class ProductStorageDto : Dto
     {
-        public long ProductId { get; set; }
-        public long StorageId { get; set; }
+        public int ProductId { get; set; }
+        public int StorageId { get; set; }
         public int AmountOfProduct { get; set; }
     }
 }
