+++ CommercialApplication/ApplicationLayer/Validation/ProductStorage/ProductStorageDeleteValidator.cs
@@ -26,7 +26,7 @@ namespace CommercialApplicationCommand.ApplicationLayer.Validation.ProductStorag
                 .WithMessage("Storage name specified doesn't exist in the database");
         }
 
-        private bool ValidateStorageId(long id)
+        private bool ValidateStorageId(int id)
         {
             using (NpgsqlConnection connection = this.databaseConnectionFactory.Instance.Create())
             {
@@ -34,7 +34,7 @@ namespace CommercialApplicationCommand.ApplicationLayer.Validation.ProductStorag
             }
         }
 
-        private bool ValidateProductId(long id)
+        private bool ValidateProductId(int id)
         {
             using (NpgsqlConnection connection = this.databaseConnectionFactory.Instance.Create())
             {
